requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dj-database-url==0.2.1
Django==4.1.7
django-cors-headers==3.14.0
django-crispy-forms==1.11.2
django-environ==0.10.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-translate==3.11.0
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
joblib==1.2.0
nltk==3.8.1
proto-plus==1.22.2
protobuf==4.22.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
Wikipedia-API==0.5.8
